StatComponent에 아래 변수들 추가.
m_CurMoveSpeed;
m_CurAdditionalMoveSpeed;
m_CurAttackSpeed;
m_CurAdditionalAttackSpeed;

아이템이나 스킬같은걸로 인한 속도증가시 실제 캐릭터의 속도에 반영하기 위함.

+++ Source/HyobinsProject/private/Utility/Utility.cpp
@@ -65,7 +65,7 @@ void Utility::InitAttackInformations(const FString& assetPath, TMap<FName, FAtta
 			if (crowdControlEnum != nullptr)
 			{
 				const int32 index = crowdControlEnum->GetIndexByName(data.crowdControlType);
-				crowdControlType = ECrowdControlType((uint8)index);
+				crowdControlType = ECrowdControlType(index);
 			}
 
 			FAttackInfoStruct attackInfoStruct;
